# apps/shared/fastapi_docs.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "coding-agent backend",
    "version": "0.1.0"
  },
  "paths": {
    "/tasks/{task_id}": {
      "get": {
        "summary": "Get Task",
        "description": "返回任务状态（含生命周期 status、上下文窗口占用与派生 execution_status）。\n\n参数:\n    task_id: 来自路由的任务标识。\n    task_service: 通过依赖注入的任务 service。\n\n返回:\n    ``TaskResponse``：已存储的任务状态（含 ``status``/``execution_status``，\n    以及 ``context_usage_used`` 与动态计算的 ``context_window_total``）。\n\n异常:\n    HTTPException: 当任务不存在时抛出。\n\n副作用:\n    无。",
        "operationId": "get_task_tasks__task_id__get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Task",
        "description": "删除单个任务及其级联的轮次与运行时事件。\n\n参数:\n    task_id: 来自路由的任务标识。\n    task_service: 通过依赖注入的任务 service（负责级联清理）。\n\n返回:\n    删除结果 ``DeleteTaskResponse``。\n\n异常:\n    HTTPException: 当任务不存在时抛出。\n\n副作用:\n    级联删除该任务下的轮次、运行时事件与任务自身。",
        "operationId": "delete_task_tasks__task_id__delete",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteTaskResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/turns": {
      "get": {
        "summary": "List Turns",
        "description": "列出某任务下的全部 turn（支撑多轮历史展示）。\n\n参数:\n    task_id: 来自路由的任务标识。\n    turn_service: 通过依赖注入的轮次 service。\n\n返回:\n    该任务下按创建时间升序的 ``TurnResponse`` 列表。\n\n异常:\n    HTTPException: 当任务不存在（级联 KeyError）时抛出。\n\n副作用:\n    无。",
        "operationId": "list_turns_tasks__task_id__turns_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TurnResponse"
                  },
                  "title": "Response List Turns Tasks  Task Id  Turns Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Turn",
        "description": "为已有任务追加一个 pending 轮次。\n\n该端点只负责把用户本轮输入持久化为一个新的、处于 ``pending`` 状态的\n``Turn`` 记录，并立即返回其 ``turn_id``；**不会**触发 Agent 运行。\n真正的执行由客户端随后对 ``GET /turns/{turn_id}/stream`` 的订阅触发，\n这样客户端可以先拿到 ``turn_id`` 再建立 SSE 连接，避免竞态。\n\n参数:\n    task_id: 来自路由的任务标识。\n    payload: 包含本轮用户输入文本的请求体（``input_text``）。\n    turn_service: 通过依赖注入的轮次 service。\n\n返回:\n    创建后的 ``TurnResponse``，包含 ``turn_id``、所属 ``task_id``、状态与\n    创建时间等字段，供客户端建立后续 SSE 流使用。\n\n异常:\n    HTTPException: 当任务不存在（404）、模型不支持视觉（422）或输入非法（400）时抛出。\n\n副作用:\n    在存储中创建 turn 记录，但不启动运行、不产生运行时事件。",
        "operationId": "create_turn_tasks__task_id__turns_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTurnRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TurnResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/events": {
      "get": {
        "summary": "Replay Task Events",
        "description": "回放某任务下的完整运行时事件流（按 turn + sequence 升序）。\n\n用于打开任务时一次性重建含思考 / 工具调用 / 状态变更的 timeline，支撑历史\n回看与刷新后重连重渲染；只读查询，不重新执行 Agent。\n\n参数:\n    task_id: 来自路由的任务标识。\n    task_service: 通过依赖注入的任务 service（用于任务存在性守卫）。\n    event_service: 通过依赖注入的运行时事件 service。\n\n返回:\n    按 ``(turn_id, sequence)`` 升序排列的事件回放列表，无记录时返回空列表。\n\n异常:\n    HTTPException: 当任务不存在时抛出（沿用任务级 404 守卫）。\n\n副作用:\n    无（只读查询）。",
        "operationId": "replay_task_events_tasks__task_id__events_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuntimeEventResponse"
                  },
                  "title": "Response Replay Task Events Tasks  Task Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/children": {
      "get": {
        "summary": "List Child Tasks",
        "description": "列出某任务下的全部子任务（委派子任务）。\n\n子任务不出现在工作区对话列表，但可通过父任务的该端点下钻查看其轨迹。\n\n参数:\n    task_id: 来自路由的父任务标识。\n    task_service: 通过依赖注入的任务 service。\n\n返回:\n    该父任务的直接子任务 ``TaskResponse`` 列表；无子任务时返回空列表。\n\n异常:\n    HTTPException: 当父任务不存在（级联 KeyError）时抛出。\n\n副作用:\n    无（只读查询）。",
        "operationId": "list_child_tasks_tasks__task_id__children_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Task Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/TaskResponse"
                  },
                  "title": "Response List Child Tasks Tasks  Task Id  Children Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/tasks/{task_id}/turns/{turn_id}/events": {
      "get": {
        "summary": "Replay Turn Events",
        "description": "回放某轮次下的运行时事件流（按 sequence 升序）。\n\n用于单轮详情页 / 单轮重连重渲染；只读查询，不重新执行 Agent。该轮次需\n隶属于路由中的 ``task_id``（交由调用方保证一致性，本端点不重复校验归属）。\n\n参数:\n    task_id: 来自路由的任务标识（仅作为 URL 层级语义锚点）。\n    turn_id: 来自路由的轮次标识。\n    event_service: 通过依赖注入的运行时事件 service。\n\n返回:\n    按 ``sequence`` 升序排列的该轮次事件回放列表，无记录时返回空列表。\n\n异常:\n    无。\n\n副作用:\n    无（只读查询）。",
        "operationId": "replay_turn_events_tasks__task_id__turns__turn_id__events_get",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "turn_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Turn Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RuntimeEventResponse"
                  },
                  "title": "Response Replay Turn Events Tasks  Task Id  Turns  Turn Id  Events Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AttachmentRef": {
        "properties": {
          "kind": {
            "type": "string",
            "enum": [
              "image",
              "file",
              "directory",
              "url"
            ],
            "title": "Kind"
          },
          "ref": {
            "type": "string",
            "title": "Ref"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "ref"
        ],
        "title": "AttachmentRef"
      },
...
      "CreateTurnRequest": {
        "properties": {
          "input_text": {
            "type": "string",
            "title": "Input Text"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "provider_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider Id"
          },
          "attachments": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/AttachmentRef"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Attachments"
          },
          "reasoning_effort": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Effort"
          }
        },
        "type": "object",
        "required": [
          "input_text"
        ],
        "title": "CreateTurnRequest",
        "description": "校验轮次创建请求体。\n\n参数:\n    input_text: 非空的本轮用户输入。\n    provider_id: 可选，模型归属的厂商标识（指向 ``providers.id``）；None 表示未选厂商。\n    model_name: 可选，本 turn 关联的模型名；None 表示未选模型。\n    attachments: 可选，本 turn 携带的附件引用列表（结构化，含类型与引用）。\n        仅做结构合法性校验；workspace 边界、视觉能力等业务规则在 service 层判定。\n    reasoning_effort: 可选，本 turn 思考努力等级；None 表示 max。该值经\n        ``turn_service.create_turn`` 透传落库到 ``turns.reasoning_effort``。\n\n返回:\n    Pydantic 请求模型。\n\n异常:\n    ValueError: 当输入为空白或附件列表非法时抛出。\n\n副作用:\n    无。本层只做结构校验，不访问文件系统、不做业务规则判定。"
      },
...
      "DeleteTaskResponse": {
        "properties": {
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "deleted": {
            "type": "boolean",
            "title": "Deleted"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "deleted"
        ],
        "title": "DeleteTaskResponse",
        "description": "校验并序列化任务删除结果响应。\n\n参数:\n    task_id: 被删除的任务标识。\n    deleted: 是否删除成功。\n\n返回:\n    Pydantic 响应模型。\n\n异常:\n    无。\n\n副作用:\n    无。"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RuntimeEventResponse": {
        "properties": {
          "event_id": {
            "type": "string",
            "title": "Event Id"
          },
          "event_type": {
            "type": "string",
            "title": "Event Type"
          },
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "turn_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Turn Id"
          },
          "sequence": {
            "type": "integer",
            "title": "Sequence"
          },
          "payload": {
            "type": "object",
            "title": "Payload"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "message_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message Id"
          },
          "tool_call_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Tool Call Id"
          }
        },
        "type": "object",
        "required": [
          "event_id",
          "event_type",
          "task_id",
          "sequence",
          "payload",
          "created_at"
        ],
        "title": "RuntimeEventResponse",
        "description": "回放用运行时事件响应模型（与持久化事件字典对齐）。\n\n字段对齐 ``RuntimeEventCrud.list_by_task`` / ``list_by_turn`` 返回的事件字典，\n即可直接序列化回前端 ``RuntimeEvent``（思考 / 工具调用 / 状态变更）的回放帧。\n\n参数:\n    event_id: 唯一事件标识。\n    event_type: 稳定的机器可读事件类型字符串。\n    task_id: 关联任务标识。\n    turn_id: 关联轮次标识。\n    sequence: 事件在单次运行流内的排序号。\n    payload: 与 ``event_type`` 匹配的事件负载字典。\n    created_at: 事件创建时间的 UTC 文本。\n    message_id: 可选用户可读消息标识。\n    tool_call_id: 可选工具调用标识。\n\n返回:\n    Pydantic 响应模型。\n\n异常:\n    无。\n\n副作用:\n    无。"
      },
      "TaskResponse": {
        "properties": {
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "workspace_id": {
            "type": "integer",
            "title": "Workspace Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "execution_status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Execution Status"
          },
          "task_type": {
            "type": "string",
            "title": "Task Type",
            "default": "user"
          },
          "parent_task_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Task Id"
          },
          "parent_turn_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Turn Id"
          },
          "delegation_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Delegation Id"
          },
          "context_usage_used": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Usage Used"
          },
          "context_window_total": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Context Window Total"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "task_id",
          "workspace_id",
          "title",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "TaskResponse",
        "description": "校验并序列化任务状态响应（与 ``TaskRecord.to_dict()`` 对齐）。\n\n任务不再绑定 agent：本响应不含 agent_id，agent 维度由 turn 维度承载。\n\n参数:\n    task_id: 任务标识。\n    workspace_id: 所属工作区标识。\n    title: 任务标题。\n    status: 任务生命周期状态。\n    execution_status: 派生执行状态，可能为 None。\n    task_type: 任务类型，``\"user\"`` 为用户创建，``\"delegation\"`` 为委派子任务。\n    parent_task_id: 父任务标识，仅委派子任务有值。\n    parent_turn_id: 父轮次标识，仅委派子任务有值。\n    delegation_id: 所属委派标识，仅委派子任务有值。\n    context_usage_used: 最近一次上下文窗口已用 token（运行时回写，可能为 None）。\n    context_window_total: 该任务模型的上下文窗口上限 token（由 resolve_context_window\n        动态计算，可能为 None）。\n    created_at: 创建时间文本。\n    updated_at: 更新时间文本。\n\n返回:\n    Pydantic 响应模型。\n\n异常:\n    无。\n\n副作用:\n    无。"
      },
      "TurnResponse": {
        "properties": {
          "turn_id": {
            "type": "integer",
            "title": "Turn Id"
          },
          "task_id": {
            "type": "integer",
            "title": "Task Id"
          },
          "input_text": {
            "type": "string",
            "title": "Input Text"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "end_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Reason"
          },
          "response_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Text"
          },
          "model_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Name"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "turn_id",
          "task_id",
          "input_text",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "TurnResponse",
        "description": "校验并序列化轮次状态响应（与 ``TurnRecord`` 字段对齐）。\n\n参数:\n    turn_id: 轮次标识。\n    task_id: 所属任务标识。\n    input_text: 本轮用户输入。\n    status: 轮次状态。\n    end_reason: 终态原因，可能为 None。\n    response_text: Agent 回复文本，可能为 None。\n    agent_id: 驱动该轮次的 agent 标识，可能为 None。\n    model_name: 该轮实际所用模型名（解析后落库，D1）；历史 turn 或运行期尚未\n        回写的轮次为 None（不表示 Auto，仅表示未落库）。\n    created_at: 创建时间文本。\n    updated_at: 更新时间文本。\n\n返回:\n    Pydantic 响应模型。\n\n异常:\n    无。\n\n副作用:\n    无。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}